Initial commit with the little helper tool to-do-timeline
Regarding the used applications file- and code-structure,
(my favourite of all times), you can find details 
and a template of it on github:
https://github.com/Reinerth/Application-Template-MVC-best-Practice

Regarding the application:
A to-do-list based on JavaScript, php and a txt-file.
There is no database integrated. (means, without SQL).
Yet, tested only with edge-browser.

Ment to be a little helper-tool, 
mainly for a developer, because a web-server and php are needed 
and the application is not safe enough for a public server.
See limitations below. 

Known limitations:
I. SINGLE-USER-application (Details see below) 
II. Not working with IIS-server (Details see below)
III. No XSS-catches (Details see below) 

Details I:
No database behind, is the reason 
why it can only be used by one single user.
If two different users, would make a change on the timeline-file,
they would overwrite each others changes.

Details II:
Not the best technical approach, but
for getting and setting the content to the file,
The POST-Method was used, because GET has limited allowed size. 
This is the reason why it is not working on IIS. 
Microsofts Server IIS (Internet Information Services) 
does not accept requesting a file via POST. 
HTTPD-server from Apache (e.g. XAMPP) accepts POST to request a file.

Details III:
It would be quite easy to break the tool if trying.
There are no integrated security-issues: 
- no checks of XSS-injections
- no validity-checks on input-fields
- no lock- or error-catches on the amount of tasks
(I tried 1000 and all went fine - surely not ment to be for more)